python/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

boto3==1.34.4
botocore==1.34.4
cloudpathlib==0.17.0
imageio==2.33.1
ipympl==0.9.3
ipython==8.18.1
ipywidgets==8.1.1
json5==0.9.14
matplotlib==3.8.2
numpy==1.26.2
Pillow==10.1.0
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
sshconf==0.2.5
torchvision==0.16.2
tqdm==4.66.1
